(mostly) fixed sample accurate grain timing
Added grain speed on birth, grains now inherit velocity from sphere. Still getting clicking on grains smaller than ~200, which is limiting the possibilities of the synthesis model. Also timing is not perfect. Not sure that Unity likes creating multiple new game objects every update. Might need to move back to a model with a statically sized array of grain gameobjects based on max grains. Maybe it would be more efficient to simply recreate the grain script? or just refresh it or something.