Fixed collision/emitter grain fighting and moved to lookup table for windowing
Grain timing issues 95% fixed, good enough! Changed windowing back to lookup table instead of real-time calculation to decrease processing time, however the mapping function to align samples with the table still uses heaps of CPU. Would be good to have a solution for this, as I think it's the biggest consumer at this point. Since the grains are arbitrarily sized, I'm not sure if there's another option... so many samples!! :|